trigger gen refactoring
